add macros
