+++ .gitignore
@@ -13,3 +13,4 @@ __pycache__/
 
 main.py
 Inputs/test_unfolded.py
+test_main.py
